+++ source/main.c
@@ -62,7 +62,7 @@ static void PrintHeader(OSScreenID bufferNum)
     OSScreenPutFontEx(bufferNum, -4, 0, " __  __ _ _                 _ _   _  __      ___ _   _   _ ");
     OSScreenPutFontEx(bufferNum, -4, 1, "|  \\/  (_|_)___ ___ _ _  __| | | | | \\ \\    / (_|_) | | | |");
     OSScreenPutFontEx(bufferNum, -4, 2, "| |\\/| | | (_-</ -_) ' \\/ _` | |_| |  \\ \\/\\/ /| | | | |_| |");
-    OSScreenPutFontEx(bufferNum, -4, 3, "|_|  |_|_|_/__/\\___|_||_\\__,_|\\___/    \\_/\\_/ |_|_|  \\___/  v1.2.0");
+    OSScreenPutFontEx(bufferNum, -4, 3, "|_|  |_|_|_/__/\\___|_||_\\__,_|\\___/    \\_/\\_/ |_|_|  \\___/  v1.3.0");
 }
 
 /**